pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "n8n-cli"
version = "0.1.0"
description = "AI-friendly CLI for n8n: node-level workflow ops + summarized execution data"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Moonward" }]
keywords = ["n8n", "cli", "workflow", "automation", "ai", "claude", "llm"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development",
    "Topic :: Utilities",
]
dependencies = [
    "typer>=0.12,<1",
    "httpx>=0.27,<1",
    "pydantic[email]>=2.6,<3",
    "jsonpath-ng>=1.6,<2",
    "PyYAML>=6,<7",
    "rich>=13,<15",
    "platformdirs>=4,<5",
    "python-dotenv>=1,<2",
]

[project.urls]
Homepage = "https://github.com/Yarrroo/n8n-cli"
Repository = "https://github.com/Yarrroo/n8n-cli"
Issues = "https://github.com/Yarrroo/n8n-cli/issues"

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "pytest-cov>=5",
    "pytest-vcr>=1.0.2",
    "vcrpy>=6",
    "mypy>=1.10",
    "ruff>=0.6",
    "pre-commit>=3.7",
    "datamodel-code-generator>=0.26",
    "types-PyYAML",
]

[project.scripts]
n8n-cli = "n8n_cli.main:run"

[tool.hatch.build.targets.wheel]
packages = ["n8n_cli"]

[tool.hatch.build.targets.sdist]
include = [
    "n8n_cli/",
    "tests/",
    "README.md",
    "LICENSE",
    "CHANGELOG.md",
    "pyproject.toml",
]
exclude = [
    "tests/**/cassettes/*.yaml",
    "tests/**/cassettes/*.yml",
    "**/__pycache__",
    "**/.DS_Store",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "E", "F", "W",       # pycodestyle + pyflakes
    "I",                 # isort
    "UP",                # pyupgrade
    "B",                 # bugbear
    "SIM",               # simplify
    "RUF",               # ruff-specific
    "N",                 # pep8-naming
    "C4",                # comprehensions
    "PTH",               # use pathlib
    "TID",               # tidy imports
]
ignore = [
    "E501",  # line too long — handled by formatter
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["N", "S101"]
"n8n_cli/models/_generated/**/*.py" = ["ALL"]

[tool.ruff.format]
quote-style = "double"

[tool.mypy]
python_version = "3.11"
strict = true
plugins = ["pydantic.mypy"]
exclude = ["n8n_cli/models/_generated/", "build/", "dist/"]

[[tool.mypy.overrides]]
module = ["jsonpath_ng.*", "vcr.*"]
ignore_missing_imports = true

[tool.pydantic-mypy]
init_forbid_extra = false
init_typed = true
warn_required_dynamic_aliases = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra --strict-markers"
markers = [
    "integration: hits real n8n instance (requires N8N_CLI_TEST_INSTANCE)",
    "acceptance: full success-criteria checks",
]

[tool.coverage.run]
source = ["n8n_cli"]
omit = ["n8n_cli/models/_generated/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
]
